LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_removed_pen-expected.txt

child
done

FAIL pointer events from pen received before/after child removal at pointerdown assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerup@parent,pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-removed),pointerup@parent,pointerout@child,pointerleave@child"
FAIL pointer events from pen received before/after child removal at pointerup assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerdown@parent,pointerup@parent,pointerout@parent,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-removed),pointerout@child,pointerleave@child"
FAIL mouse events from pen received before/after child removal at pointerdown assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,(child-removed),mouseup@parent,mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,(child-removed),mousedown@child,mouseup@parent,mouseout@child,mouseleave@child"
FAIL mouse events from pen received before/after child removal at pointerup assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mousedown@parent,mouseup@parent,mouseout@parent,mouseleave@parent" but got "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-removed),mouseup@child,mouseout@child,mouseleave@child"
